requirements.txt
Flask==2.3.3
line-bot-sdk==3.21.0
google-genai==1.49.0
gunicorn==20.1.0
python-dotenv==1.0.0